db np

+++ src/lstm.R
@@ -34,7 +34,7 @@ lstm <- function(data, outcome, covariates, t_end, window_size, out_activation, 
   py$loss_fn <- loss_fn
   py$lr <- 0.001
   py$dr <- 0.5
-  py$nb_batches <- as.integer(64)
+  py$nb_batches <- as.integer(128)
   py$patience <- as.integer(2)
   py$t_end <- as.integer(t_end + 1)
   py$window_size <- as.integer(window_size)
